IOS-6: Закончил редактирование данных
