--- VKNews/VKNews/Helpers/Extensions/String+Extensions.swift
@@ -1,23 +1,0 @@
-//
-//  String+Extensions.swift
-//  VKNews
-//
-//  Created by Dmitriy Permyakov on 03.12.2024.
-//
-
-import Foundation
-
-extension String {
-
-    var formatDate: String? {
-        let dateString = self
-        let isoFormatter = ISO8601DateFormatter()
-        guard let date = isoFormatter.date(from: dateString) else {
-            return nil
-        }
-        let dateFormatter = DateFormatter()
-        dateFormatter.locale = Locale(identifier: "ru_RU")
-        dateFormatter.dateFormat = "d MMM в HH:mm"
-        return dateFormatter.string(from: date)
-    }
-}
