+++ VKNews/VKNews/Helpers/Extensions/UIKit/UINavigationController+Extensions.swift
@@ -1,0 +1,16 @@
+//
+//  UINavigationController+Extensions.swift
+//  VKNews
+//
+//  Created by Dmitriy Permyakov on 04.12.2024.
+//
+
+import UIKit
+
+extension UINavigationController {
+
+    open override func viewDidLoad() {
+        super.viewDidLoad()
+        interactivePopGestureRecognizer?.delegate = nil
+    }
+}
