+++ VKNews/VKNews/Streams/CommonModels/ImageState.swift
